script_generateQuestion/requirements.txt
# pymongo
PyPDF2
# openai
google-api-python-client
# google-cloud-storage
# functions-framework==3.*


functions-framework==3.*
pymongo==4.6.3
openai
google-cloud-storage==2.16.0